refactor: dissolve HeaderConverter onto KPF0 + the shared base
HeaderConverter bundled WMKO->EPRV conversion (only used by KPF0.to_kpf1) with
EPRV PRIMARY validation (used by KPF1.to_kpf2 / KPF2.to_kpf4), hiding the
ownership split: L0 holds WMKO-native data and owns the conversion; L1/L2/L4 are
EPRV(-like) and share the validation. Redistribute by ownership and dissolve the
class; headers.py becomes a pure reference-data module.

- KPF0 (level0.py): owns wmko_to_eprv() and build_instrument_header(), reading
  the raw L0 PRIMARY from self; to_kpf1 calls them.
- KPFDataModel (base.py): validate_eprv_primary(header, level) is now a shared
  @staticmethod; KPF1.to_kpf2 / KPF2.to_kpf4 call self.validate_eprv_primary().
- headers.py: HeaderConverter deleted; all module-level reference data
  (HEADER_MAP, EPRV/required keyword sets, WMKO_PRIMARY_KEYS, _STRUCTURAL_KEYS,
  _NUMORDER, _DRPSTATU_DEFAULT, loaders) stays and is imported by the methods.

Fix (one deliberate behavior change): build_instrument_header now returns a
fits.Header copy of the raw PRIMARY, so INSTRUMENT_HEADER preserves the raw
WMKO comments (and commentary cards) instead of flattening to scalars. Verified
the comment survives both in memory and through to_fits -> from_fits.

Out of scope (follow-up): calling validation at to_kpf1 and any new L0-native
format checks. No import cycle (base -> headers -> kpfpipe is one-directional).
Full suite green (997 passed); ruff clean.
